+++ lib/SHA.ml
@@ -35,51 +35,37 @@ module type S = sig
   module Map: Misc.Map with type key = t
 end
 
-type sha_t = { raw    : string;
-               padded : bool;   (* for hex of odd length *)
-             }
+type sha_t = {
+  raw   : string;
+  padded: bool;   (* for hex of odd length *)
+}
 
 exception Ambiguous
 
 let get_upper c = (Char.code c) land 0xf0
 
 let sha_compare x y =
-  (*Log.debugf "sha_compare: %s vs %s" (sha_to_string x) (sha_to_string y);*)
   let nx = String.length x.raw in
   let ny = String.length y.raw in
-  let res =
-    if nx = ny && not x.padded && not y.padded then
-      String.compare x.raw y.raw
-    else begin
-      let len = min nx ny in
-      let rec scan i =
-        if i = len then
-          raise Ambiguous
-        else
-        if (x.padded && y.padded) || i < len then
-          let x0 = x.raw.[i] in
-          let y0 = y.raw.[i] in
-          if x0 < y0 then
-            -1
-          else if x0 > y0 then
-            1
-          else
-            scan (i + 1)
-        else
-          let x0 = get_upper x.raw.[i] in
-          let y0 = get_upper y.raw.[i] in
-          if x0 < y0 then
-            -1
-          else if x0 > y0 then
-            1
-          else
-            raise Ambiguous
-      in
-      scan 0
-    end
-  in
-  (*Log.debugf "sha_compare: result=%d" res;*)
-  res
+  if nx = ny && not x.padded && not y.padded then String.compare x.raw y.raw
+  else
+    let len = min nx ny in
+    let rec scan i =
+      if i = len then raise Ambiguous
+      else if (x.padded && y.padded) || i < len then
+        let x0 = x.raw.[i] in
+        let y0 = y.raw.[i] in
+        if x0 < y0 then -1
+        else if x0 > y0 then 1
+        else scan (i + 1)
+      else
+        let x0 = get_upper x.raw.[i] in
+        let y0 = get_upper y.raw.[i] in
+        if x0 < y0 then -1
+        else if x0 > y0 then 1
+        else raise Ambiguous
+    in
+    scan 0
 
 module SHA1_String = struct
 
@@ -87,49 +73,31 @@ module SHA1_String = struct
 
   let length x = (* 0 <= length <= 40 *)
     let n = (String.length x.raw) * 2 in
-    if x.padded then
-      n - 1
-    else
-      n
+    if x.padded then n - 1 else n
 
-  let is_short x = (length x) < 40
-
-  let equal x y = (x=y)
-
+  let is_short x = length x < 40
+  let equal x y = x.padded = y.padded && String.compare x.raw y.raw = 0
   let hash x = Hashtbl.hash x.raw
-
   let compare = sha_compare
 
   let is_prefix p x =
     let np = length p in
     let nx = length x in
-    if np > nx then
-      false
-    else if np = nx then
-      equal p x
+    if np > nx then false
+    else if np = nx then equal p x
     else
       let n =
-        if p.padded then
-          (String.length p.raw) - 1
-        else
-          String.length p.raw
+        if p.padded then String.length p.raw - 1
+        else String.length p.raw
       in
       try
-        for i = 0 to n - 1 do
-          if p.raw.[i] <> x.raw.[i] then
-            raise Exit
-        done;
-        if p.padded then
-          (get_upper p.raw.[n]) = (get_upper x.raw.[n])
-        else
-          true
+        for i = 0 to n - 1 do if p.raw.[i] <> x.raw.[i] then raise Exit done;
+        if p.padded then get_upper p.raw.[n] = get_upper x.raw.[n] else true
       with
         Exit -> false
 
   let lt x y = compare x y < 0
-
   let to_raw x = x.raw
-
   let of_raw x = { raw=x; padded=false; }
 
   let of_string str =
@@ -145,42 +113,22 @@ module SHA1_String = struct
 
   let to_hex t =
     let `Hex h = Hex.of_string t.raw in
-    if t.padded then
-      String.sub h 0 ((String.length h) - 1)
-    else
-      h
+    if t.padded then String.sub h 0 ((String.length h) - 1)
+    else h
 
   let of_hex h =
     let len = String.length h in
     let to_be_padded = (len mod 2) = 1 in
-    let h' =
-      if to_be_padded then
-        h ^ "0"
-      else
-        h
-    in
-    { raw    = Hex.to_string (`Hex h');
-      padded = to_be_padded;
-    }
+    let h' = if to_be_padded then h ^ "0" else h in
+    { raw = Hex.to_string (`Hex h'); padded = to_be_padded; }
 
-  let zero =
-    of_hex (String.make 40 '0')
-
+  let zero = of_hex (String.make 40 '0')
   let pretty = to_hex
-
-  let pp_hum ppf t = Format.fprintf ppf "%s" (pretty t)
-
-  let input buf =
-    { raw=Mstruct.get_string buf 20; padded=false; }
-
-  let add buf ?level:_ t =
-    Buffer.add_string buf t.raw
-
-  let input_hex buf =
-    of_hex (Mstruct.get_string buf (Mstruct.length buf))
-
-  let add_hex buf t =
-    Buffer.add_string buf (to_hex t)
+  let pp ppf t = Format.fprintf ppf "%s" (pretty t)
+  let input buf = { raw=Mstruct.get_string buf 20; padded=false; }
+  let add buf ?level:_ t = Buffer.add_string buf t.raw
+  let input_hex buf = of_hex (Mstruct.get_string buf (Mstruct.length buf))
+  let add_hex buf t = Buffer.add_string buf (to_hex t)
 
   module X = struct
     type t = sha_t