+++ lib/memory.ml
@@ -26,7 +26,7 @@ type t = {
   root   : string;
   level  : int;
   values : (SHA.t, Value.t) Hashtbl.t;
-  refs   : (Reference.t, SHA.Commit.t) Hashtbl.t;
+  refs   : (Reference.t, [`S of SHA.Commit.t | `R of Reference.t]) Hashtbl.t;
   mutable head : Reference.head_contents option;
 }
 
@@ -34,12 +34,12 @@ let root t = t.root
 let level t = t.level
 
 let stores = Hashtbl.create 1024
+let default_root = "root"
+let clear ?(root=default_root) () = Hashtbl.remove stores root
+let clear_all () = Hashtbl.reset stores
 
-let create ?root ?(level=6) () =
+let create ?(root=default_root) ?(level=6) () =
   if level < 0 || level > 9 then failwith "level should be between 0 and 9";
-  let root = match root with
-    | None   -> "root"
-    | Some r -> r in
   let t =
     try Hashtbl.find stores root
     with Not_found ->
@@ -52,10 +52,6 @@ let create ?root ?(level=6) () =
       Hashtbl.add stores root t;
       t in
   Lwt.return t
-
-let clear t =
-  Hashtbl.remove stores t.root;
-  Lwt.return_unit
 
 let write t value =
   let inflated = Misc.with_buffer (fun buf -> Value.add_inflated buf value) in
@@ -128,9 +124,12 @@ let references t =
 let mem_reference t ref =
   Lwt.return (Hashtbl.mem t.refs ref)
 
-let read_reference t ref =
+let rec read_reference t ref =
   Log.info "Reading %s" (Reference.pretty ref);
-  try Lwt.return (Some (Hashtbl.find t.refs ref))
+  try
+    match Hashtbl.find t.refs ref with
+    | `S s -> Lwt.return (Some s)
+    | `R r -> read_reference t r
   with Not_found -> Lwt.return_none
 
 let read_head t =
@@ -153,7 +152,7 @@ let write_head t c =
 
 let write_reference t ref sha1 =
   Log.info "Writing %s" (Reference.pretty ref);
-  Hashtbl.replace t.refs ref sha1;
+  Hashtbl.replace t.refs ref (`S sha1);
   Lwt.return_unit
 
 let read_index _t = Lwt.return Index.empty