+++ lib/packed_refs.ml
@@ -27,7 +27,7 @@ let of_line line =
   else if line.[0] = '#' then
     let str = String.sub line 1 (String.length line - 1) in
     Some (`Comment str)
-  else match Misc.string_lsplit2 line ~on:' ' with
+  else match Stringext.cut line ~on:" " with
     | None  -> None
     | Some (sha1, ref) ->
       let sha1 = SHA.Commit.of_hex sha1 in
@@ -48,8 +48,8 @@ let add buf ?level:_ t =
   let ppf = Format.formatter_of_buffer buf in
   List.iter (to_line ppf) t
 
-let pp_hum ppf t = List.iter (to_line ppf) t
-let pretty = Misc.pretty pp_hum
+let pp ppf t = List.iter (to_line ppf) t
+let pretty = Misc.pretty pp
 
 let input buf =
   let rec aux acc =