+++ lib/reference.mli
@@ -38,10 +38,19 @@ type head_contents =
   | Ref of t
 (** The possible HEAD contents. *)
 
-val head_contents: SHA.Commit.t Map.t -> SHA.Commit.t -> head_contents
-(** Compute the head contents. The result is either the hex
-    representation of the SHA or something like {i ref: <ref>} if the
-    SHA has already a reference pointing to it. *)
+val head_contents_of_string: string -> head_contents
+(** Parse the contents of HEAD *)
+
+val head_contents_of_commit: SHA.Commit.t Map.t -> SHA.Commit.t -> head_contents
+(** Build a head contents from a commit hash. The result is either the
+    hex representation of the SHA or something like {i ref: <ref>} if
+    the SHA has already a reference pointing to it. *)
+
+val pp_head_contents: Format.formatter -> head_contents -> unit
+(** Pretty-print head contents. *)
+
+val equal_head_contents: head_contents -> head_contents -> bool
+(** Compare head contents. *)
 
 val is_valid: t -> bool
 (** Check if a reference can be stored on disk properly. *)